fix(structure): refactored the mock and shared stuff, added a components directory.
