+++ src/app/components/markdown-control/markdown-control.component.scss
@@ -1,4 +1,4 @@
-@import '../../assets/stylesheets/base';
+@import '../../../assets/stylesheets/base';
 
 .markdown-mark {
   width: 20px;
