+++ src/app/header/DummySpace.service.ts
@@ -1,9 +1,9 @@
-import { HttpService } from './../shared/http-service';
+import { HttpService } from '../services/http-service';
 import { WIT_API_URL, Space } from 'ngx-fabric8-wit';
 import { Injectable, Component, Inject } from '@angular/core';
 import { Headers, Http } from '@angular/http';
 
-import { MockHttp } from '../shared/mock-http';
+import { MockHttp } from '../mock/mock-http';
 
 @Injectable()
 export class DummySpace {
