+++ src/app/mock/standalone/spaces.service.factory.ts
