+++ src/app/mock/standalone/spaces.service.ts
