+++ src/app/toolbar-panel/toolbar-panel.component.ts
@@ -12,7 +12,7 @@ import { AreaModel } from '../models/area.model';
 import { AreaService } from '../services/area.service';
 import { FilterModel } from './../models/filter.model';
 import { CollaboratorService } from '../services/collaborator.service';
-import { FilterService } from '../shared/filter.service';
+import { FilterService } from '../services/filter.service';
 import { WorkItemService } from '../services/work-item.service';
 import { WorkItemListEntryComponent } from './../work-item/work-item-list/work-item-list-entry/work-item-list-entry.component';
 import { WorkItemType } from './../models/work-item-type';
