+++ src/app/work-item/work-item-list/planner-list.module.ts
@@ -35,8 +35,8 @@ import { WorkItemListComponent } from './work-item-list.component';
 import { WorkItemListEntryComponent } from './work-item-list-entry/work-item-list-entry.component';
 import { WorkItemQuickAddModule } from '../work-item-quick-add/work-item-quick-add.module';
 import { WorkItemService } from '../../services/work-item.service';
-import { MockHttp } from './../../shared/mock-http';
-import { HttpService } from './../../shared/http-service';
+import { MockHttp } from '../../mock/mock-http';
+import { HttpService } from '../../services/http-service';
 
 let providers = [];
